Move HomeRetaliationCard mapping to specific file
